CCS matrix padding constraints removed (#41)
I added padding with dummy constraints
[here](https://github.com/NethermindEth/latticefold/pull/37). This
introduced memory problems with large witnesses. I now realised the
padding (mentioned in Definition 4.3) is redundant - the problem of
having `Mz_mles` of the correct size to be able to evaluate `r` is
actually handled already
[here](https://github.com/NethermindEth/latticefold/blob/7980b88f81ad82e678b531950a903f5ff0b35a34/latticefold/src/utils/mle.rs#L89).

So it suffices just to set `m` correctly when instantiating the
constraint matrix, the padding is redundant and is removed in this PR.

Without padding, there is no problem with huge witnesses (for example
for witness of size `2**14`, with padding we got `2**14 x 2**14` matrix,
while without padding it's only `num_of_constraints x 2**14`.