Compute the NTT/iNTT only once (#69)
Adding `f_coeff`, the coefficient representation, to the witness. Since
we need to compute it anyway let's compute it at the very initialisation
of a witness struct. So that we can easily access it without
recomputing. In particular, `decompose_B_vec_into_k_vec` would recompute
the coefficient form of the witness.