Jolt-based sumcheck prove round impl (#116)
Jolt-based Sumcheck implementation. Adapted to prove single rounds,
following our current logic. Also adapted to run serialized.
Can be employed by enabling the `jolt-sumcheck` feature.
Provides a `comb_func` to use other MLE combination schemes, other than
the current product-based implementation.

`comb_func` is currently employed as an argument for
`prove_as_subprotocol` if the feature `jolt-sumcheck` is enabled. While
it could be associated with `VirtualPolynomial`, adding it there
requires dynamic dispatch (or an additional template) which adds a bit
of complexity.
`comb_func` also uses as an argument the `ProverState` which may provide
the user access to some values which evolve throughout the Sumcheck
rounds.
If `comb_func` proves to be useful, we can think of a more suitable
interface for it.

Compared to the current Sumcheck implementation, this one performs 1-5%
better serialized and 1-4% worse parallelized.

* This PR  also absorbs lattirust `VirtualPolynomial`.