Move to Jolt-based sumcheck (#129)
Moves from our original product-based sumcheck implementation, to Jolt's
`comb_fn`-based implementation.

Handling the `comb_fn` is still a bit finicky. I've further explored
moving it to the `VirtualPolynomial` as a field, however using either
`dyn`amic dispatch or being `<`templated`>` is a bit awkward, mainly
because we are using it as a closure. However using it as a closure is
useful for us due to their versatility and advantage in capturing stuff.
Currently, as it stands provides the lowest impact in readability and
performance. Let me know if you guys have any other ideas on how to
handle this.

Also removes `VirtualPolynomial`.